SDC Manager Changes
