--- docs/resources/sdc_expansion.md
@@ -1,105 +1,0 @@
----
-# Copyright (c) 2023 Dell Inc., or its subsidiaries. All Rights Reserved.
-# 
-# Licensed under the Mozilla Public License Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-# 
-#     http://mozilla.org/MPL/2.0/
-# 
-# 
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-title: "powerflex_sdc_expansion resource"
-linkTitle: "powerflex_sdc_expansion"
-page_title: "powerflex_sdc_expansion Resource - powerflex"
-subcategory: ""
-description: |-
-  This resource can be used to add the SDC in PowerFlex Cluster.
----
-
-# powerflex_sdc_expansion (Resource)
-
-This resource can be used to add the SDC in PowerFlex Cluster.
-
-
-## Example Usage
-
-```terraform
-# Command to run this tf file : terraform init && terraform plan && terraform apply
-# Create, Update, Delete is supported for this resource
-resource "powerflex_sdc_expansion" "test" {
-  mdm_password = "Password"
-  lia_password = "Password"
-  cluster_details = [
-    {
-      ip               = "IP"
-      username         = "Username"
-      password         = "Password"
-      operating_system = "linux"
-      is_mdm_or_tb     = "Primary"
-      is_sdc           = "Yes"
-    },
-    {
-      ip               = "IP"
-      username         = "Username"
-      password         = "Password"
-      operating_system = "linux"
-      is_mdm_or_tb     = "Secondary"
-      is_sdc           = "Yes"
-    },
-    {
-      ip               = "IP"
-      username         = "Username"
-      password         = "Password"
-      operating_system = "linux"
-      is_mdm_or_tb     = "TB"
-      is_sdc           = "Yes"
-    },
-    {
-      ip               = "IP"
-      username         = "Username"
-      password         = "Password"
-      operating_system = "linux"
-      is_mdm_or_tb     = "Standby"
-      is_sdc           = "Yes"
-    },
-  ]
-}
-```
-
-<!-- schema generated by tfplugindocs -->
-## Schema
-
-### Required
-
-- `cluster_details` (Attributes Set) List of SDC Expansion Server Details. (see [below for nested schema](#nestedatt--cluster_details))
-- `lia_password` (String, Sensitive) LIA Password to connect MDM Server.
-- `mdm_password` (String, Sensitive) MDM Password to connect MDM Server.
-
-### Read-Only
-
-- `id` (String) The ID of the package.
-- `installed_sdc_ips` (String) List of installed SDC IPs
-
-<a id="nestedatt--cluster_details"></a>
-### Nested Schema for `cluster_details`
-
-Required:
-
-- `ip` (String) IP of the node
-- `is_sdc` (String) whether this node is SDC or not,The acceptable value is `Yes` or `No`.
-- `operating_system` (String) Operating System on the node
-- `password` (String, Sensitive) Password of the node
-
-Optional:
-
-- `is_mdm_or_tb` (String) Whether this works as MDM or Tie Breaker,The acceptable value is `Primary`, `Secondary`, `TB`, `Standby` or blank. Default value is blank
-- `performance_profile` (String) Performance Profile of SDC, The acceptable value is `High` or `Compact`.
-- `sdc_name` (String) Name of the SDC
-- `username` (String) Username of the node
-
