+++ docs/resources/sdc_manager.md
@@ -1,0 +1,76 @@
+---
+# Copyright (c) 2023 Dell Inc., or its subsidiaries. All Rights Reserved.
+# 
+# Licensed under the Mozilla Public License Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+# 
+#     http://mozilla.org/MPL/2.0/
+# 
+# 
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+title: "powerflex_sdc_manager resource"
+linkTitle: "powerflex_sdc_manager"
+page_title: "powerflex_sdc_manager Resource - powerflex"
+subcategory: ""
+description: |-
+  This resource can be used to add the SDC in PowerFlex Cluster.
+---
+
+# powerflex_sdc_manager (Resource)
+
+This resource can be used to add the SDC in PowerFlex Cluster.
+
+
+
+
+<!-- schema generated by tfplugindocs -->
+## Schema
+
+### Optional
+
+- `id` (String) ID of the SDC to manage. This can be retrieved from the PowerFlex website. Cannot be updated.
+- `lia_password` (String, Sensitive) LIA Password to connect MDM Server.
+- `mdm_password` (String, Sensitive) MDM Password to connect MDM Server.
+- `name` (String) Name of the SDC to manage.
+- `sdc_details` (Attributes Set) List of SDC Expansion Server Details. (see [below for nested schema](#nestedatt--sdc_details))
+
+<a id="nestedatt--sdc_details"></a>
+### Nested Schema for `sdc_details`
+
+Optional:
+
+- `id` (String) ID of the SDC to manage. This can be retrieved from the PowerFlex website. Cannot be updated.
+- `ip` (String) IP of the node
+- `is_mdm_or_tb` (String) Whether this works as MDM or Tie Breaker,The acceptable value is `Primary`, `Secondary`, `TB`, `Standby` or blank. Default value is blank
+- `is_sdc` (String) whether this node is SDC or not,The acceptable value is `Yes` or `No`.
+- `name` (String) Name of the SDC to manage.
+- `operating_system` (String) Operating System on the node
+- `password` (String, Sensitive) Password of the node
+- `performance_profile` (String) Performance Profile of SDC, The acceptable value is `High` or `Compact`.
+- `username` (String) Username of the node
+
+Read-Only:
+
+- `last_updated` (String) The Last updated timestamp of the SDC.
+- `links` (Attributes List) The Links of the fetched SDC. (see [below for nested schema](#nestedatt--sdc_details--links))
+- `mdm_connection_state` (String) The MDM connection status of the fetched SDC.
+- `on_vmware` (Boolean) If the fetched SDC is on vmware.
+- `sdc_approved` (Boolean) If the fetched SDC is approved.
+- `sdc_guid` (String) The GUID of the fetched SDC.
+- `sdc_ip` (String) The IP of the fetched SDC.
+- `system_id` (String) The System ID of the fetched SDC.
+
+<a id="nestedatt--sdc_details--links"></a>
+### Nested Schema for `sdc_details.links`
+
+Read-Only:
+
+- `href` (String) The Links-HREF of the fetched SDC.
+- `rel` (String) The Links-Rel of the fetched SDC.
+
